+++ internal/dataplane/kongstate/route_test.go
@@ -364,7 +364,7 @@ func TestNormalizeProtocols(t *testing.T) {
 
 	assert.NotPanics(func() {
 		var nilUpstream *Upstream
-		nilUpstream.override(nil, nil)
+		nilUpstream.override(nil)
 	})
 }