+++ internal/store/store_test.go
@@ -69,7 +69,6 @@ spec:
 	t.Log("verifying the integrity of the cache store")
 	assert.Len(t, cs.IngressV1.List(), 1)
 	assert.Len(t, cs.Service.List(), 1)
-	assert.Len(t, cs.KongIngress.List(), 0)
 	_, exists, err = cs.Get(&corev1.Service{ObjectMeta: metav1.ObjectMeta{Namespace: "doesntexist", Name: "doesntexist"}})
 	assert.NoError(t, err)
 	assert.False(t, exists)